CVProject/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.3
cssselect2==0.8.0
distro==1.9.0
Django==5.2.1
django-environ==0.12.0
django-xhtml2pdf==0.0.4
djangorestframework==3.16.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
kombu==5.5.4
lxml==5.4.0
openai==1.86.0
oscrypto==1.3.0
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
pypdf==5.6.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==6.2.0
reportlab==4.4.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.17
